% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{CG2FBgn}
\alias{CG2FBgn}
\title{Returns a data_frame for mapping FB CG numbers to FBgn}
\usage{
CG2FBgn()
}
\value{
A mapping data_frame
}
\description{
Returns a data_frame for mapping FB CG numbers to FBgn
}
\keyword{annotation}

