% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hts_util.R
\name{do.jaccard}
\alias{do.jaccard}
\title{Performs a Jaccard comparison between two bed files}
\usage{
do.jaccard(a, b)
}
\arguments{
\item{a}{first bed file}

\item{b}{second bed file}
}
\value{
A vector containing Jaccard distances
}
\description{
Generates Jaccard simlirity of two bed files ( a measure of similarity, less sensitive to noise than Spearman, Pearson, etc.
Requires bedtools to be installed and in environment PATH
}
\keyword{chip-seq}
\keyword{jaccard}

