% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{fbToCommon}
\alias{fbToCommon}
\title{Returns a mapping data_frame from org.Dm.eg.db}
\usage{
fbToCommon()
}
\value{
A mapping data_frame
}
\description{
Mapping data.frame of FBgn <-> entrez <-> Gene Description
}
\keyword{annotation}

