% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illumina.barcodes}
\alias{illumina.barcodes}
\title{A list of Illumina TruSeq barcodes, accurate as of 10/2014}
\format{A vector of illumina barcodes. Position corresponds to barcode number}
\usage{
illumina.barcodes
}
\description{
A list of Illumina TruSeq barcodes, accurate as of 10/2014
}
\keyword{datasets}

