% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{import.macs2.narrowpeak}
\alias{import.macs2.narrowpeak}
\title{Import a MACS2 NarrowPeak file as a GRanges object}
\usage{
import.macs2.narrowpeak(file)
}
\arguments{
\item{file}{Path to a NarrowPeak file for import}
}
\value{
A GRanges object
}
\description{
Creates a GRanges object from the NarrowPeak file, taking into account switch from zero-base to one-base nomenclature. Adds and labels data as metadata.
}
\keyword{GRanges}
\keyword{MACS2}
\keyword{chip}

