% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chip.R
\name{import.macs2.xls}
\alias{import.macs2.xls}
\title{Reads a MACS2 .xls file into a GRanges, preserving metadata}
\usage{
import.macs2.xls(file)
}
\arguments{
\item{file}{}
}
\value{
g A GRange
}
\description{
Reads a MACS2 .xls file into a GRAnges object with metadata
}
\keyword{GRanges}
\keyword{MACS2}
\keyword{chip}

