% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hts_util.R
\name{spear.matrix}
\alias{spear.matrix}
\title{Generates a matrix of spearman coefficients}
\usage{
spear.matrix(df)
}
\arguments{
\item{df}{A data frame, each column being a set of data to correlate}
}
\value{
A matrix of spearman coefficients
}
\description{
Generates a matrix of spearman coefficients
Recommend plotting using a heatmap
}
\details{
library(gplots)
library(RColorBrewer)
heatmap.2(coeffs,Rowv=T,Colv=T,symm=T,key=T,trace="none",col=brewer.pal(9,"Purples"))
}
\keyword{Spearman}

